new stylesheet

+++ src/main/java/me/wmn/service/PackageServiceImpl.java
@@ -1,8 +1,11 @@
 package me.wmn.service;
 
+import java.io.File;
 import java.util.List;
 
+import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import me.wmn.domain.OSPackage;
@@ -11,10 +14,13 @@ import me.wmn.persistence.IPackageDao;
 @Service
 public class PackageServiceImpl implements IPackageService{
 	
-	private final String PACKAGE_NAME_PATTERN = "%s_%s_%s_%i.%s";
+	private final String PACKAGE_NAME_PATTERN = "%s_%s_%s_%d.%s";
 	
 	@Autowired
 	private IPackageDao packageDao;
+	
+	@Value("${attachement.folder}")
+	private String uploadFolder;
 	
 	@Override
 	public void addPackage(OSPackage osp) {
@@ -23,7 +29,11 @@ public class PackageServiceImpl implements IPackageService{
 
 	@Override
 	public void deleteById(int id) {
-		this.packageDao.deleteById(id);
+		OSPackage osp = this.getById(id);
+		if(osp != null){
+			FileUtils.deleteQuietly(new File(this.uploadFolder + osp.getPackageName()));
+			this.packageDao.deleteById(id);
+		}
 	}
 
 	public IPackageDao getPackageDao() {
@@ -41,14 +51,20 @@ public class PackageServiceImpl implements IPackageService{
 
 	@Override
 	public String getPackageName(String productName, String versionType, String versionName, int build, String suffix) {
-		return String.format(this.PACKAGE_NAME_PATTERN, productName, versionName, build, suffix);
+		return String.format(this.PACKAGE_NAME_PATTERN, productName,versionType, versionName, build, suffix);
 	}
 	
 	public OSPackage getById(int id){
 		return this.packageDao.getById(id);
 	}
-	
-	
 
+	public String getUploadFolder() {
+		return uploadFolder;
+	}
+
+	public void setUploadFolder(String uploadFolder) {
+		this.uploadFolder = uploadFolder;
+	}
+	
 	
 }
